Configure docker-compose to pull SOLR_URL from env
In order to set different SOLR_URLs via env variables, docker-compose
has to pass the env variable along to the container. This adds that,
plus it adds `.env` to .gitignore so that a .env file can be used to set
environment variables. If the env variable isn't present (as in
production) then the SOLR_URL is pulled from the credentials file as
normal.