Disable autocomplete
This is related to #10.

I'm disabling autocomplete for now until I have a chance to look into
why it isn't working. I thought I had configured it for our Solr index
but it has never actually worked; our Solr version is a bit old (4.X) so
the autocomplete implementation may be too outdated.