Create separate "access" and "location" facets
This is in reference to #5.

The ticket mentions that the Access facet should have two values: Online
Copy and In Library Copy. We do have a subset of records that have a
physical copy and an online copy, and the way the bib_location_codes and
item_location_codes Solr fields are implemented, and the way our
location codes are set up, it's very easy to find things that have at
least one copy that's online, but not the other way around. Ultimately
I'm not sure we even need an option to limit to the physical copy, so
I'm going to see how this looks to everyone with just the one option.